D: 第一版资源管理， Shader Texture Material Model用ResouceHandle记录生命周期，运行时创建也在ResourceManager中创建。

+++ Runtime/Core/Object.h
@@ -7,7 +7,7 @@ namespace EngineCore
     {
     public:
         InstanceID id;
-        inline const uint32_t GetInstanceID() const{ return id.v;};
+        inline const uint64_t GetInstanceID() const{ return id.v;};
 
         Object(){ id = InstanceIDGenerator::New();};
     };
